+++ hashtreedb/query_replace.go
@@ -3,10 +3,10 @@ package hashtreedb
 /**
  *
  */
-func (ins *QueryInstance) replace(searchitem *FindValueOffsetItem, valuedatas []byte, SaveValueSegmentOffset uint32) (ValueSegmentOffset uint32, err error) {
+func (ins *QueryInstance) replace(searchitem *FindValueOffsetItem, valuedatas []byte, SaveValueSegmentOffset int64) (ValueSegmentOffset uint32, err error) {
 	if valuedatas != nil {
 		return ins.writeSegmentData(searchitem.ValueSegmentOffset, valuedatas)
 	} else {
-		return SaveValueSegmentOffset, nil // do not really write
+		return uint32(SaveValueSegmentOffset), nil // do not really write
 	}
 }
